# backend/document/swagger.yaml
openapi: 3.0.3
info:
  title: EnviroConnect
  description: API documentation for IMA EnviroConnect
  version: 1.0.0
servers:
  - url: http://103.200.20.117:5000
  - url: http://localhost:5000
...
paths:
...
  /api/users/{userid}/added-activity:
    put:
      tags:
        - users
      parameters:
        - in: path
          name: userid
          description: id of user
          required: true
          schema:
            type: string
      summary: add 1 activity into user when user register a activity
      description: add 1 activity into user when user register a activity
      operationId: registerActivity
      requestBody:
        required: true
        description: activityid need to add
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterActivityRequest"
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterActivityResponse"
        400:
          description: Invalid input
  /api/users/{userid}/done-activity:
    put:
      tags:
        - users
      parameters:
        - in: path
          name: userid
          description: id of user
          required: true
          schema:
            type: string
      summary: mark done 1 activity
      description: mark done 1 activity
      operationId: MarkDoneActivity
      requestBody:
        required: true
        description: activityid need to mark done
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkDoneActivityRequest"
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarkDoneActivityResponse"
        400:
          description: Invalid input
...
components:
  schemas:
...
    RegisterActivityResponse:
      type: object
      properties:
        _id:
          type: string
          example: 652274823b7da4dfe80a963e
        name:
          type: string
          example: Huynh Ngoc Nhu
        email:
          type: string
          example: nhuhuynh@gmail.com
        address:
          type: string
          example: 244A Lê Văn Quới, Bình Trị Đông, Bình Tân, Thành phố Hồ Chí Minh
        location:
          type: object
          properties:
            lat:
              type: number
              example: 10.7812029
            long:
              type: number
              example: 106.6030883
        isOrganizer:
          type: boolean
          example: false
        buddy:
          type: array
          example: []
        activities:
          type: array
          items:
            type: object
            properties:
              activityId:
                type: string
              joinDate:
                type: string
              status:
                type: number
            example:
              - activityId: 651d42afa4e4a95676215752,
                joinDate: 2023-10-20T03:24:35.996Z,
                status: 0
              - activityId: 651d42a12345a95676215752,
                joinDate: 2023-10-16T03:24:35.996Z,
                status: 0
    RegisterActivityRequest:
      type: object
      properties:
        activityid:
          type: string
          example: 651d42a12345a95676215752
    MarkDoneActivityRequest:
      type: object
      properties:
        activityid:
          type: string
          example: 651d42a12345a95676215752
    MarkDoneActivityResponse:
      type: object
      properties:
        _id:
          type: string
          example: 652274823b7da4dfe80a963e
        name:
          type: string
          example: Huynh Ngoc Nhu
        email:
          type: string
          example: nhuhuynh@gmail.com
        address:
          type: string
          example: 244A Lê Văn Quới, Bình Trị Đông, Bình Tân, Thành phố Hồ Chí Minh
        location:
          type: object
          properties:
            lat:
              type: number
              example: 10.7812029
            long:
              type: number
              example: 106.6030883
        isOrganizer:
          type: boolean
          example: false
        buddy:
          type: array
          example: []
        activities:
          type: array
          items:
            type: object
            properties:
              activityId:
                type: string
              joinDate:
                type: string
              status:
                type: number
            example:
              - activityId: 651d42afa4e4a95676215752,
                joinDate: 2023-10-20T03:24:35.996Z,
                status: 0
              - activityId: 651d42a12345a95676215752,
                joinDate: 2023-10-16T03:24:35.996Z,
                status: 1
...